@@ -1,7 +1,7 @@
 package com.visualpathit.account.model;
 
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 import java.io.Serializable;
 import java.util.Set;
@@ -44,7 +44,7 @@ public class User implements Serializable {
     private Set<Role> roles;
     /** {@inheritDoc}} !*/
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     /** {@link User#id} */
     public Long getId() {
         return id;
