@@ -42,6 +42,6 @@ public class UserServiceImpl implements UserService {
     }
     @Override
 	public User findById(long id){
-    	return userRepository.findOne(id);
+    	return userRepository.findById(id);
     }
 }
