@@ -6,15 +6,15 @@ import org.springframework.web.servlet.view.InternalResourceViewResolver;
 
 public class StandaloneMvcTestViewResolver extends InternalResourceViewResolver {
 
-    public StandaloneMvcTestViewResolver() {
-        super();
-    }
+	public StandaloneMvcTestViewResolver() {
+		super();
+	}
 
-    @Override
-    protected AbstractUrlBasedView buildView(final String viewName) throws Exception {
-        final InternalResourceView view = (InternalResourceView) super.buildView(viewName);
-        // prevent checking for circular view paths
-        view.setPreventDispatchLoop(false);
-        return view;
-    }
+	@Override
+	protected AbstractUrlBasedView buildView(final String viewName) throws Exception {
+		final InternalResourceView view = (InternalResourceView) super.buildView(viewName);
+		// prevent checking for circular view paths
+		view.setPreventDispatchLoop(false);
+		return view;
+	}
 }
